Move filter function to helper
