+++ controllers/keyword_test.go
@@ -252,52 +252,3 @@ func TestUploadKeywordWithGuestUser(t *testing.T) {
 	assert.Equal(t, http.StatusFound, response.Code)
 	assert.Equal(t, "/login", response.Header().Get("Location"))
 }
-
-func TestFilterValidConditionsWithValidQueryString(t *testing.T) {
-	queryString := map[string][]string{
-		"filter[keyword]": {"test"},
-	}
-
-	result := filterValidConditions(queryString)
-
-	expected := []models.Condition{
-		{
-			ConditionName: "keyword",
-			Value:         "test",
-		},
-	}
-
-	assert.Equal(t, expected, result)
-}
-
-func TestFilterValidConditionsWithoutQueryString(t *testing.T) {
-	result := filterValidConditions(nil)
-
-	var expected []models.Condition
-
-	assert.Equal(t, expected, result)
-}
-
-func TestFilterValidConditionsWithInvalidQueryString(t *testing.T) {
-	queryString := map[string][]string{
-		"filter[invalid]": {"test"},
-	}
-
-	result := filterValidConditions(queryString)
-
-	var expected []models.Condition
-
-	assert.Equal(t, expected, result)
-}
-
-func TestFilterValidConditionsWithBlankQueryStringValue(t *testing.T) {
-	queryString := map[string][]string{
-		"filter[keyword]": {""},
-	}
-
-	result := filterValidConditions(queryString)
-
-	var expected []models.Condition
-
-	assert.Equal(t, expected, result)
-}
